% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app3.R
\name{crop_spd_app}
\alias{crop_spd_app}
\title{Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.}
\usage{
crop_spd_app(
  spdat,
  shapeind = 1,
  tableind = 1,
  cropview_name = "pick",
  shape_feature_name = "location_id",
  table_feature_id = "cell_id"
)
}
\arguments{
\item{spdat}{instance of SpatialData}

\item{shapeind}{numeric(1), index of shapes(spdat) to use as base display}

\item{tableind}{numeric(1) index of tables(spdat) to subset according to cropping}

\item{cropview_name}{character(1) name used for updating list elements; if already present,
a message is given indicating overwriting will occur.}

\item{shape_feature_name}{character(1) string used to identify shape element identifiers in ShapeFrame,
defaults to "location_id"}

\item{table_feature_id}{character(1) strings used to identify element of colData that can be matched
by values of shape_feature_name in ShapeFrame of cropped SpatialData}
}
\description{
Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.
}
\examples{
brdat = SpatialData.data::Breast2fov_10x()
blk = crop_spd_app(brdat, shape_feature_name = "__index_level_0__")
blk
}
