% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app5.R
\name{crop_spd_simple}
\alias{crop_spd_simple}
\title{Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.}
\usage{
crop_spd_simple(spdatname)
}
\arguments{
\item{spdat}{instance of SpatialData}
}
\description{
Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.
}
\examples{
blk = crop_spd_simple("Breast2fov_10x")
blk
}
