% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app7.R
\name{crop_spd_simply}
\alias{crop_spd_simply}
\title{Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.}
\usage{
crop_spd_simply(spdatname, typetag = "celltype_major")
}
\arguments{
\item{spdatname}{character(1) name of function that retrieves SpatialData instance
defined in SpatialData.data package}

\item{typetag}{character(1) name of colData variable in table() that maps cells to type, when available}
}
\description{
Use clicks on shape display of a SpatialData instance to define
a cropping region, produce a new ShapeFrame, visualize it, and
update the shapes and tables elements to respect the selected
shape elements.
}
\examples{
blk = crop_spd_simply("JanesickBreastXeniumRep1")
blk
}
